backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
python-dotenv>=1.0.1
slowapi>=0.1.9
google-generativeai>=0.5.4
pytest>=8.2.0
pytest-asyncio>=0.23.6
